Google Maps API and Database Implementation
1st version

+++ Galeria2/src/com/example/Galeria2/ImageTask.java
@@ -9,10 +9,10 @@ public class ImageTask {
     private String name;
     private String address;
     private Float rating;
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
 
-    public ImageTask(String name, String address, Float rating, Float latitude, Float longitude) {
+    public ImageTask(String name, String address, Float rating, Double latitude, Double longitude) {
         this.name = name;
         this.address = address;
         this.rating = rating;
@@ -44,19 +44,19 @@ public class ImageTask {
         this.rating = rating;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
